Added all kinds of things to EmnImaging, which is now basically a playground and could use some cleaning up.
Also added parsting of line_annot.txt, and parsing and generation of .words files.

Also added a simple fsharp project for testing these interactively.

Added a whole bunch of image processing primitives, badly implemented, inside the EmnImageTestDisplay.csproj project, but actually the whole of imaging needs to be refactored to support multiple dimensions, and FFTW and mapping of unmanaged memory.  This is too slow.